requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit
langchain
langchain_core
langchain_community
langchain-ollama
langchain_text_splitters
faiss-cpu
beautifulsoup4
html2text
numpy
fitz
pymupdf
google-generativeai